rename main directory
