+++ proc_chirp.py
@@ -1,7 +1,7 @@
 from pyspecdata import *
 from scipy.optimize import minimize,leastsq
-from proc_scripts import *
-from proc_scripts import postproc_dict
+from pyspecProcScripts import *
+from pyspecProcScripts import postproc_dict
 from sympy import symbols
 from scipy.signal import tukey
 from pylab import *
