+++ pyspecProcScripts/slice_FID_from_echo.py
