+++ pyspecProcScripts/third_level/process_IR.py
