+++ test_integration.py
@@ -1,8 +1,8 @@
 from pyspecdata import *
 from pylab import *
 from matplotlib import *
-from proc_scripts import *
-from proc_scripts.correlation_alignment_ODNP import correl_align
+from pyspecProcScripts import *
+from pyspecProcScripts.correlation_alignment_ODNP import correl_align
 import numpy as np
 fl = figlist_var()
 signal_pathway = {'ph1': 1, 'ph2':0}
